Refactor stationary energy review helpers
