+++ climate-advisor/service/app/services/stationary_energy/__init__.py
@@ -1,17 +1,1 @@
 """Stationary Energy service package."""
-
-from app.services.stationary_energy.stationary_energy_draft_service import (
-    COMMIT_ACCEPTED_CAPABILITY,
-    LOAD_CONTEXT_CAPABILITY,
-    StationaryEnergyDraftService,
-)
-from app.services.stationary_energy.stationary_energy_proposal_builder import (
-    build_deterministic_proposals,
-)
-
-__all__ = [
-    "COMMIT_ACCEPTED_CAPABILITY",
-    "LOAD_CONTEXT_CAPABILITY",
-    "StationaryEnergyDraftService",
-    "build_deterministic_proposals",
-]
